refactor(customer): split CustomerDashboard (509→132) em hook + componentes
Extrai verbatim de src/components/CustomerDashboard.tsx:
- useCustomerDashboard: gamification score, checagem de endereços e derivados
  (displayName, ferramentas urgentes, ação prioritária, partição de pedidos)
- priority.ts (computePriority), DashboardHero, PriorityCard, GamificationMini,
  PedidosAndamento, FerramentasAtencao, AcoesRapidas, OrderRow,
  types.ts, config.ts (statusConfig/stagger/fadeUp/QUICK_ACTIONS)
- +20 testes (lógica de prioridade + presentacionais)

Wrappers framer-motion mantidos no componente pai (orquestração do stagger
preservada). Comportamento 1:1. tsc baseline=0, eslint=0.
